test:closable modal & expandable table row tests (#635)
* fix(modal): closable default true

* fix(table): revert

* fix(modal): true if undefined

* test(modal): toBeHidden

* test(table): synchronized test

---------
