+++ packages/core/src/components/modal/test/closable-false/modal.e2e.ts-snapshots/x-button-is-hidden-1-linux.png
