+++ packages/core/src/components/table/table/test/expandable-row-methods/methods.e2e.ts-snapshots/tds-table-expandable-row-methods-click-collapse-button-1-linux.png
